requirements.txt
orgo
google-generativeai
python-dotenv
requests
beautifulsoup4
rich 